Fundamental fixes
Offset added for player pivot to be relative to feet.
Offset added for tile pivot to be relative to front edge of the surface of the tile.
can center player on tile
update player z value depending on the current tile they are on (to determine if they are infront or behind)
other improvements
code messy, need to start refactoring, unit testing and xml commenting