working on terrain bounds
more complicated than anticipated. added another tilemap; one to hold colliders which are offset above the tile, and one to hold colliders with no offset (for infront of and behind the terrain). need to work on corners which will need dynamic colliders (colliders which change depending on players position on terrain).